--- galini_dashboard/API/pointpack04/_sym.json
@@ -1,1 +1,0 @@
-["{\"src\": \"0_1_0_1\", \"dst\": \"0_0_0_1_1\", \"variations\": [[[1, 7], [2, 1], [5, 2], [6, 5], [7, 6]], [[1, 7], [2, 1], [5, 6], [6, 5], [7, 2]], [[1, 7], [2, 5], [5, 2], [6, 1], [7, 6]], [[1, 7], [2, 5], [5, 6], [6, 1], [7, 2]]]}", "{\"src\": \"0_0_0_1_1\", \"dst\": \"0_1_0_1\", \"variations\": [[[1, 2], [2, 5], [5, 6], [6, 7], [7, 1]], [[1, 2], [2, 7], [5, 6], [6, 5], [7, 1]], [[1, 6], [2, 5], [5, 2], [6, 7], [7, 1]], [[1, 6], [2, 7], [5, 2], [6, 5], [7, 1]]]}"]
